Issue #52: Add unit test for config, and move speech configuration.
