+++ sksurgerybard/algorithms/bard_config_algorithms.py
@@ -4,17 +4,11 @@
 
 import os
 import glob
-from sys import modules
-from threading import Thread
 import numpy as np
 from sksurgerycore.configuration.configuration_manager import \
         ConfigurationManager
 from sksurgerybard.algorithms.interaction import BardKBEvent, \
         BardMouseEvent, BardFootSwitchEvent
-try:
-    from sksurgerybard.algorithms.speech_interaction import BardSpeechInteractor
-except ModuleNotFoundError:
-    pass
 
 
 def get_calibration_filenames(calibration_dir):
@@ -151,7 +145,7 @@ def configure_model_and_ref(model_config):
         reference2model = np.loadtxt(reference2model_file)
 
     return ref_data, reference2model, models_path, visible_anatomy, \
-                    smoothing_buffer
+        smoothing_buffer
 
 
 def configure_pointer(pointer_config):
@@ -247,31 +241,3 @@ def configure_interaction(interaction_config, vtk_window, pointer_writer,
     if interaction_config.get('mouse', False):
         vtk_window.AddObserver("LeftButtonPressEvent",
                                BardMouseEvent(bard_visualisation))
-
-
-def configure_speech_interaction(speech_config, bard_visualisation):
-    """
-    Configures a BARD speech interactor, creating a thread for it to run in
-    :param: The speech configuration, which is passed to the speech
-        interaction service.
-    :param: A visualisation manager to be triggered by speech events
-    :raises: KeyError if speech not configured.
-    :raises: ModuleNotError if sksurgeryspeech is not installed.
-    :return: the speechinteractor, which needs to be stopped by the
-    calling application, with speech_int.stop_listener()
-    """
-    if not speech_config:
-        raise KeyError("Requested speech interaction without" +
-                       " speech config key")
-
-    if 'sksurgeryspeech' not in modules:
-        raise ModuleNotFoundError(
-            "Requested speech interaction without " +
-            "sksurgeryspeech installed check your setup.")
-
-    speech_int = BardSpeechInteractor(speech_config,
-                                      bard_visualisation)
-    speech_thread = Thread(target=speech_int, daemon=True)
-    speech_thread.start()
-
-    return speech_int
